CPsched.Backend/requirements.txt
attr==0.3.2
Cython
fastapi[standard]
beautifulsoup4==4.12.3
chromedriver_autoinstaller==0.6.4
colorama==0.4.6
ConfigParser==7.1.0
cryptography==42.0.8
docutils==0.21.2
fastapi==0.115.6
filelock==3.16.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
keyring==25.6.0
Pillow==11.0.0
protobuf==5.29.2
pyOpenSSL==24.3.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
selenium==4.27.1
Sphinx==8.1.3
thread==2.0.5
urllib3_secure_extra==0.1.0

